+++ src/main/java/club/inq/team1/service/impl/NaverMapServiceImpl.java
@@ -1,7 +1,8 @@
-package club.inq.team1.client;
+package club.inq.team1.service.impl;
 
 import club.inq.team1.dto.response.GeocodeResponseDTO;
 import club.inq.team1.dto.response.ReverseGeocodeResponseDTO;
+import club.inq.team1.service.MapService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
@@ -13,7 +14,7 @@ import org.springframework.web.util.UriComponentsBuilder;
  */
 
 @Service
-public class NaverMapService implements MapService {
+public class NaverMapServiceImpl implements MapService {
 
     @Value("${naver.map.client-id}")
     private String clientId;
