+++ src/test/java/club/inq/team1/client/NaverMapServiceImplTest.java
@@ -2,16 +2,17 @@ package club.inq.team1.client;
 
 import club.inq.team1.dto.response.GeocodeResponseDTO;
 import club.inq.team1.dto.response.ReverseGeocodeResponseDTO;
+import club.inq.team1.service.impl.NaverMapServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class NaverMapServiceTest {
+class NaverMapServiceImplTest {
 
     @Autowired
-    NaverMapService naverMapClient;
+    NaverMapServiceImpl naverMapClient;
 
     GeocodeResponseDTO geocodeResponseDTO ;
     ReverseGeocodeResponseDTO reverseGeocodeResponseDTO ;
